itmo_web/requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.0
Django==4.1.5
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
phonenumbers==8.13.4
Pillow==9.4.0
pycparser==2.21
PyMySQL==1.0.2
six==1.16.0
sqlparse==0.4.3
